+++ src/background.js
@@ -1,9 +1,12 @@
 "use strict";
 
+require( "file-loader?name=[name].[ext]!./manifest.json" );
+
 import Logger from './modules/logger.js';
 import Settings from './modules/settings.js';
 import Badge from './modules/badge.js';
 import Authentication from './modules/authentication.js';
+import { PocketError, PocketNotice } from './modules/constants.js';
 
 // --------------------------
 
@@ -11,15 +14,6 @@ const consumerKey = '58817-addc87503598b7ed29e5bf72';
 
 
 // --- API REQUEST HELPER ---
-
-const PocketError = {
-  GENERIC:      'generic',
-  UNREACHABLE:  'unreachable',
-  UNAUTHORIZED: 'unauthorized',
-  PERMISSIONS:  'missing_permissions',
-  RATE_LIMIT:   'user_rate_limit_reached'
-};
-
 
 function prepareRequest( url, action, successCallback, errorCallback ) {
   let request = new XMLHttpRequest();
@@ -134,12 +128,11 @@ var AuthenticationProcess = ( function() {
 
         // Retrieve the items and update the badge count
         retrieveItems( true );
-        Badge.updateCount( itemsList );
       };
 
-      browser.storage.local.get( 'requestToken', function( data ) {
+      browser.storage.local.get( 'requestToken', function( { requestToken } ) {
         let request = prepareRequest( 'https://getpocket.com/v3/oauth/authorize', 'POST', onSuccess );
-        let requestParams = JSON.stringify({ consumer_key: consumerKey, code: data.requestToken });
+        let requestParams = JSON.stringify({ consumer_key: consumerKey, code: requestToken });
         request.send( requestParams );
       });
     }
@@ -173,13 +166,13 @@ function retrieveItems( force ) {
   const intervalWithoutReload = 15*60;
   const currentTimestamp      = ( Date.now()/1000 | 0 );
 
-  browser.storage.local.get([ 'items', 'last_retrieve' ], function( data ) {
-    Logger.log( "retrieve items timeout: " + ( currentTimestamp - data.last_retrieve ) + ' / ' + intervalWithoutReload );
+  browser.storage.local.get([ 'items', 'last_retrieve' ], function( { items, last_retrieve } ) {
+    Logger.log( "retrieve items timeout: " + ( currentTimestamp - last_retrieve ) + ' / ' + intervalWithoutReload );
 
-    if ( force || !data.items || !data.last_retrieve ) {
+    if ( force || !items || !last_retrieve ) {
       // If force == true, we always reload the whole list
       retrieveFirst();
-    } else if( currentTimestamp - data.last_retrieve > intervalWithoutReload ) {
+    } else if( currentTimestamp - last_retrieve > intervalWithoutReload ) {
       // If we already have sync, check if intervalWithoutReload is past, then we can reload
       retrieveDiff();
     } else {
@@ -193,7 +186,7 @@ function retrieveItems( force ) {
 function retrieveFirst() {
   Logger.log('(retrieve first)');
 
-  browser.storage.local.get('access_token', function( data ) {
+  browser.storage.local.get('access_token', function( { access_token } ) {
     let onSuccess = function( response ) {
       Logger.log(Object.keys( response.list ).length + ' items in the response');
 
@@ -201,12 +194,12 @@ function retrieveFirst() {
       for( let itemId in response.list ) {
         let item = response.list[ itemId ];
 
-        // NOTE: in some cases, resolved_title/resolved_url do not exist so I then
-        //       fallback to given_url (there's no given_url equivalent for the title)
+        // https://getpocket.com/developer/docs/v3/retrieve
+        // given_url should be used if the user wants to view the item.
         itemsList.push({
           id:             item.item_id,
-          resolved_title: item.resolved_title,
-          resolved_url:   item.resolved_url || item.given_url,
+          resolved_title: item.given_title || item.resolved_title,
+          resolved_url:   item.given_url || item.resolved_url,
           created_at:     item.time_added
         });
       };
@@ -222,12 +215,14 @@ function retrieveFirst() {
       // TODO: Do this once in the "retrieveItems" method
       chrome.runtime.sendMessage({ action: 'retrieved-items' });
 
+      // Updates the tabs page actions
+      redrawPageActionAllTabs();
     };
 
     let request = prepareRequest( 'https://getpocket.com/v3/get', 'POST', onSuccess );
     let requestParams = JSON.stringify( {
       consumer_key: consumerKey,
-      access_token: data.access_token,
+      access_token: access_token,
       detailType: 'simple',
     });
 
@@ -239,10 +234,10 @@ function retrieveFirst() {
 function retrieveDiff() {
   Logger.log('(retrieve diff)');
 
-  browser.storage.local.get( ['access_token', 'last_retrieve', 'items'], function( data ) {
+  browser.storage.local.get( ['access_token', 'last_retrieve', 'items'], function( { access_token, last_retrieve, items } ) {
     let onSuccess = function( response ) {
       Logger.log(Object.keys(response.list).length + ' items in the response');
-      let allItems = JSON.parse( data.items );
+      let allItems = JSON.parse( items );
 
       // TODO: Extract this into a dedicated method
       for( let itemId in response.list ) {
@@ -268,8 +263,8 @@ function retrieveDiff() {
               // we just update the missing fields
               Logger.log("ITEM " + itemId + "(" + item.resolved_title + ") ALREADY PRESENT, WILL BE UPDATED" );
               allItems[ itemIdx ] = Object.assign( allItems[ itemIdx ], {
-                resolved_title: item.resolved_title,
-                resolved_url:   item.resolved_url,
+                resolved_title: item.given_title || item.resolved_title,
+                resolved_url:   item.given_url || item.resolved_url,
                 created_at:     item.time_added
               });
             } else {
@@ -277,8 +272,8 @@ function retrieveDiff() {
               Logger.log("NEED TO ADD: " + itemId + ' (' + item.resolved_title + ')' );
               allItems.push({
                 id:             item.item_id,
-                resolved_title: item.resolved_title,
-                resolved_url:   item.resolved_url,
+                resolved_title: item.given_title || item.resolved_title,
+                resolved_url:   item.given_url || item.resolved_url,
                 created_at:     item.time_added
               });
             }
@@ -299,15 +294,19 @@ function retrieveDiff() {
       // Send a message back to the UI
       // TODO: Do this once in the "retrieveItems" method
       chrome.runtime.sendMessage({ action: 'retrieved-items' });
+
+
+      // Updates the tabs page actions
+      redrawPageActionAllTabs();
     };
 
     let request = prepareRequest( 'https://getpocket.com/v3/get', 'POST', onSuccess );
     let requestParams = JSON.stringify({
       consumer_key: consumerKey,
-      access_token: data.access_token,
+      access_token: access_token,
       detailType: 'simple',
       state: 'all',
-      since: data.last_retrieve
+      since: last_retrieve
     });
 
     request.send( requestParams );
@@ -317,17 +316,18 @@ function retrieveDiff() {
 
 // TODO Be sure to url-encode the parameters you are sending. Otherwise if your url or title
 //      have characters like ? or &, they will often break the request.
-function addItem( url ) {
+function addItem( url, title ) {
   Logger.log('(addItem)');
 
   browser.storage.local.get( [ 'access_token', 'items' ], function( data ) {
     let onSuccess = function( response ) {
       let itemsList = JSON.parse( data.items );
       let newItem   = response.item;
+
       itemsList.push({
         id:             newItem.item_id,
-        resolved_title: newItem.title,
-        resolved_url:   newItem.resolved_url || newItem.given_url,
+        resolved_title: title || newItem.title,
+        resolved_url:   url,
         created_at:     ( Date.now()/1000 | 0 )
       });
 
@@ -346,19 +346,37 @@ function addItem( url ) {
 
         // Send a message back to the UI
         chrome.runtime.sendMessage({ action: 'added-item', id: newItem.item_id });
+
       }, 2500);
+
+      browser.tabs.query( { url: url } ).then( function( tabs ) {
+        for( const tab of tabs ) {
+          redrawPageAction( tab.id, tab.url );
+        }
+      });
     };
 
-    let request = prepareRequest( 'https://getpocket.com/v3/add', 'POST', onSuccess );
-    let requestParams = JSON.stringify({
-      consumer_key: consumerKey,
-      access_token: data.access_token,
-      url: url
+
+    let itemsList = JSON.parse( data.items );
+    let alreadyContainsItem = itemsList.some( function( item, index, array ) {
+      return item.resolved_url == url;
     });
 
-    request.send( requestParams );
-  });
+    if( alreadyContainsItem === true ) {
+      // Instead of just logging, send an event back to the UI
+      chrome.runtime.sendMessage( { notice: PocketNotice.ALREADY_IN_LIST } );
+    } else if ( alreadyContainsItem === false ) {
+      let request = prepareRequest( 'https://getpocket.com/v3/add', 'POST', onSuccess );
+      let requestParams = JSON.stringify({
+        consumer_key: consumerKey,
+        access_token: data.access_token,
+        url: url,
+        title: title
+      });
 
+      request.send( requestParams );
+    }
+  });
 }
 
 
@@ -371,6 +389,7 @@ function markAsRead( itemId ) {
       Logger.log('onload - itemId = ' + itemId );
       let items = JSON.parse( data.items );
       let removedItemIdx = items.findIndex( function( item ) { return item.id === itemId });
+      let removedItem = items[ removedItemIdx ];
 
       if( removedItemIdx >= 0 ) {
         Logger.log('the item ' + itemId + ' has been found and removed');
@@ -385,6 +404,15 @@ function markAsRead( itemId ) {
 
       // Send a message back to the UI
       chrome.runtime.sendMessage({ action: 'marked-as-read', id: itemId });
+
+      // Redraw page action
+      if( removedItem ) {
+        browser.tabs.query( { url: removedItem.resolved_url } ).then( function( tabs ) {
+          for( const tab of tabs ) {
+            redrawPageAction( tab.id, tab.url );
+          }
+        });
+      }
     };
 
     let request = prepareRequest( 'https://getpocket.com/v3/send', 'POST', onSuccess );
@@ -401,30 +429,65 @@ function markAsRead( itemId ) {
 }
 
 
+function openRandomItem( opt = {} ) {
+  browser.storage.local.get( 'items' ).then( function( { items } ) {
+    const parsedItems = items ? JSON.parse( items ) : [];
+    const item = items[ Math.floor( Math.random() * parsedItems.length ) ];
+    opt.url = item.resolved_url;
+
+    openItem( opt );
+  });
+}
+
+
+function openItem( { newTab, url } ) {
+  let pending;
+  if( newTab == null ) {
+    pending = Settings.init().then( function() {
+      newTab = Settings.get( 'openInNewTab' );
+    });
+  } else {
+    pending = Promise.resolve();
+  }
+
+  pending.then( function() {
+    if( newTab ) {
+      browser.tabs.create( { url } );
+    } else {
+      browser.tabs.update( { url } );
+    }
+  });
+}
+
+
 // --- MESSAGES ---
 
 chrome.runtime.onMessage.addListener( function( eventData ) {
+  Logger.log( `eventData.action:${eventData.action}` );
   switch( eventData.action ) {
     case 'authenticate':
-      Logger.log("switch:authenticate");
       AuthenticationProcess.authenticate();
       break;
     case 'retrieve-items':
-      Logger.log("switch:retrieve-items");
       retrieveItems( eventData.force );
       break;
     case 'add-item':
-      Logger.log('switch:add-item');
-      addItem( eventData.url );
+      addItem( eventData.url, eventData.title );
       break;
     case 'mark-as-read':
-      Logger.log('switch:mark-as-read');
       markAsRead( eventData.id );
       break;
     case 'update-badge-count':
-      Logger.log('switch:update-badge-count');
       Badge.updateCount();
       break;
+    case 'random-item':
+      openRandomItem();
+      break;
+    case 'read-item':
+      openItem( {url: eventData.url} );
+      break;
+    default:
+      Logger.log( `Unknown action:${eventData.action}` );
   }
 });
 
@@ -434,3 +497,84 @@ chrome.runtime.onMessage.addListener( function( eventData ) {
 Authentication.isAuthenticated().then( function() {
   Badge.updateCount();
 })
+
+
+//
+// Feature: add link to Pocket from righ-click context menu
+//
+const addLinkId = 'add-link-to-pocket';
+
+chrome.contextMenus.create({
+  id: addLinkId,
+  title: 'Add to Pocket',
+  contexts: ['link', 'page']
+});
+
+
+browser.contextMenus.onClicked.addListener( function( link, tab ) {
+  if( link.menuItemId == addLinkId ) {
+    // const url = link.linkUrl || link.pageUrl;
+    if( link.linkUrl ) {
+      addItem( url );
+    } else {
+      addItem( link.pageUrl, tab.title );
+    }
+  }
+});
+
+
+//
+// Feature: add "in-pocket" indicator
+//
+
+browser.tabs.onUpdated.addListener( function( tabId, changeInfo ) {
+  if( changeInfo.url ) {
+    redrawPageAction( tabId, changeInfo.url );
+  }
+});
+
+browser.pageAction.onClicked.addListener( function( tab ) {
+  togglePageAction( tab );
+});
+
+function redrawPageAction( tabId, url ) {
+  browser.storage.local.get( "items" ).then( function( { items } ) {
+    const parsedItems = JSON.parse( items );
+    if( parsedItems.some( i => i.resolved_url == url ) ) {
+      // item is in pocket
+      browser.pageAction.setIcon( { tabId, path: "assets/icons/inmypocket-flat-blue.svg" });
+      browser.pageAction.setTitle({ tabId, title: "Mark as read" });
+    } else {
+      browser.pageAction.setIcon( { tabId, path: "assets/icons/inmypocket-flat-grey.svg" });
+      browser.pageAction.setTitle({ tabId, title: "Add to pocket" });
+    }
+    showPageAction(tabId);
+  });
+}
+
+function redrawPageActionAllTabs() {
+  browser.tabs.query( {} ).then( function( tabs ) {
+    for( const tab of tabs ) {
+      if( tab.url ) {
+        redrawPageAction( tab.id, tab.url );
+      }
+    }
+  });
+}
+
+function showPageAction( tabId ) {
+  browser.pageAction.show( tabId );
+}
+
+function togglePageAction( tab ) {
+  browser.storage.local.get( "items" ).then( function( { items } ) {
+    const parsedItems = JSON.parse( items );
+    const item = parsedItems.find( i => i.resolved_url == tab.url );
+    if( item ) {
+      // in pocket
+      markAsRead( item.id );
+    } else {
+      addItem( tab.url, tab.title );
+    }
+  });
+}