+++ src/popup/popup.js
@@ -1,14 +1,19 @@
 "use strict";
 
+require( "file-loader?name=[path][name].[ext]!./popup.html" );
+import'./popup.scss';
+
 import Logger from '../modules/logger.js';
 import Badge from '../modules/badge.js';
 import Authentication from '../modules/authentication.js';
+import { PocketError, PocketNotice } from '../modules/constants.js';
 
 
 // --- EVENTS ---
 
 let retrieveItemsButton  = document.querySelector( '.retrieve-items' );
 let addCurrentPageButton = document.querySelector( '.add-current'  );
+let readRandomItemButton = document.querySelector( '.random-item' );
 let filterItemsInput     = document.querySelector( '.filter-items' );
 let placeholderNoResults = document.querySelector( '.search-no-results' );
 let listComponent        = document.querySelector( '.list-component' );
@@ -22,8 +27,14 @@ addCurrentPageButton.addEventListener( 'click', function() {
   MainLoader.enable();
   chrome.tabs.query({ active: true, currentWindow: true }, function ( tabs ) {
     let currentUrl = tabs[ 0 ].url;
-    chrome.runtime.sendMessage({ action: 'add-item', url: currentUrl });
+    let currentTitle = tabs[ 0 ].title;
+    chrome.runtime.sendMessage({ action: 'add-item', url: currentUrl, title: currentTitle });
   });
+});
+
+readRandomItemButton.addEventListener( 'click', function() {
+  browser.runtime.sendMessage({action: 'random-item'});
+  // window.close();
 });
 
 filterItemsInput.addEventListener( 'keyup', function() {
@@ -69,6 +80,8 @@ var UI = ( function() {
     let liElement    = document.createElement('li');
     let titleContent = document.createElement('span');
     let urlContent   = document.createElement('span');
+    let actionContainer = document.createElement('div');
+    let tickAction = document.createElement('div');
     let tickElement  = document.createElement('div');
     let tickIconFont = document.createElement('i');
     let loadElement  = document.createElement('div');
@@ -76,6 +89,8 @@ var UI = ( function() {
     liElement.className    = 'item';
     titleContent.className = 'title';
     urlContent.className   = 'url';
+    actionContainer.className = 'actions-container';
+    tickAction.className = 'tick-action';
     tickIconFont.classList.add( 'icon', 'icon-ok');
     tickElement.className  = 'tick';
     loadElement.classList.add( 'loader', 'hidden' );
@@ -89,8 +104,11 @@ var UI = ( function() {
 
     urlContent.appendChild( document.createTextNode( formatUrl( item.resolved_url ) ) );
 
-    liElement.appendChild( tickElement );
-    liElement.appendChild( loadElement );
+    tickAction.appendChild( tickElement );
+    tickAction.appendChild( loadElement );
+    actionContainer.appendChild( tickAction );
+
+    liElement.appendChild( actionContainer );
     liElement.appendChild( titleContent );
     liElement.appendChild( document.createElement('br') );
     liElement.appendChild( urlContent );
@@ -104,7 +122,7 @@ var UI = ( function() {
   }
 
   function openLink( url ) {
-    browser.tabs.create({ url: url });
+    browser.runtime.sendMessage({action: 'read-item', url});
   }
 
     function focusSearchField() {
@@ -116,9 +134,9 @@ var UI = ( function() {
 
   return {
     drawList: function() {
-      browser.storage.local.get('items', function( data ) {
-        let items           = data.items ? JSON.parse( data.items ) : [];
-        let sortedItems     = items.sort( function( a, b ) { return a.created_at < b.created_at; });
+      browser.storage.local.get('items', function( { items } ) {
+        let parsedItems     = items ? JSON.parse( items ) : [];
+        let sortedItems     = parsedItems.sort( function( a, b ) { return a.created_at < b.created_at; });
         let itemsToRender   = undefined;
         let query           = document.querySelector( '.filter-items').value;
 
@@ -202,34 +220,50 @@ document.addEventListener('DOMContentLoaded', function() {
   UI.setup();
 
   chrome.runtime.onMessage.addListener( function( eventData ) {
-    if( eventData.error ) {
+    if( eventData.error || eventData.notice ) {
       MainLoader.disable();
 
-      let errorContainer = document.querySelector( '.error-overlay' );
-      let errorMessage = 'An error occurred: ';
+      let flashContainer = document.querySelector( '.flash-overlay' );
+      let flashMessage   = '';
+      let errorClass     = 'error';
+      let noticeClass    = 'notice';
 
-      switch( eventData.error ) {
-        case PocketError.UNREACHABLE:
-          errorMessage += 'could not reach the server';
-          break;
-        case PocketError.UNAUTHORIZED:
-          errorMessage += 'unauthorized, you might need to login again';
-          break;
-        case PocketError.PERMISSIONS:
-          errorMessage += 'missing permissions';
-          break;
-        case PocketError.RATE_LIMIT:
-          errorMessage += 'max number of requests reach for this hour';
-          errorMessage += ' (reset in ' + eventData.resetDelay + ')';
-          break;
+      if( eventData.error ) {
+        flashContainer.classList.add( errorClass );
+        flashMessage = 'An error occurred: ';
+
+        switch( eventData.error ) {
+          case PocketError.UNREACHABLE:
+            flashMessage += 'could not reach the server';
+            break;
+          case PocketError.UNAUTHORIZED:
+            flashMessage += 'unauthorized, you might need to login again';
+            break;
+          case PocketError.PERMISSIONS:
+            flashMessage += 'missing permissions';
+            break;
+          case PocketError.RATE_LIMIT:
+            flashMessage += 'max number of requests reach for this hour';
+            flashMessage += ' (reset in ' + eventData.resetDelay + ')';
+            break;
+        }
+      } else if( eventData.notice ) {
+        flashContainer.classList.add( noticeClass );
+
+        switch( eventData.notice ) {
+          case PocketNotice.ALREADY_IN_LIST:
+            flashMessage = 'This page is already in your Pocket :)';
+            break;
+        }
       }
 
-      errorContainer.innerHTML = errorMessage;
-      errorContainer.classList.remove( 'hidden' );
+      flashContainer.innerHTML = flashMessage;
+      flashContainer.classList.remove( 'hidden' );
 
-      // Hide the error message after 5 seconds
+      // Hide the error message after 5 seconds and reset the class list
       setTimeout( function() {
-        errorContainer.classList.add( 'hidden' );
+        flashContainer.classList.add( 'hidden' );
+        flashContainer.classList.remove( errorClass, noticeClass );
       }, 5000 );
 
     } else {